+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.d.ts
@@ -80,6 +80,9 @@ interface SohoDataGridOptions {
   /** Is the grid editable? */
   editable?: boolean;
 
+  /** Allows you to provide a function so you can set some rows to disabled based on the data and/or the row index. */
+  isRowDisabled?: SohoIsRowDisabledFunction;
+
   /** Makes a readonly "list". */
   isList?: boolean;
 
@@ -400,6 +403,11 @@ type SohoDataGridResponseFunction = (
   results: Object[],
   request: SohoDataGridSourceRequest
 ) => void;
+
+type SohoIsRowDisabledFunction = (
+  actualIndex: number,
+  rowData: any
+) => boolean;
 
 type SohoDataGridResultsTextFunction = (
   source: any,